include entry definitions within owning policy list
