+++ catalystwan/models/policy/list/as_path.py
@@ -2,8 +2,15 @@
 
 from typing import List, Literal
 
-from catalystwan.models.policy.lists_entries import ASPathListEntry
+from pydantic import BaseModel, ConfigDict, Field
+
 from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
+
+
+class ASPathListEntry(BaseModel):
+    model_config = ConfigDict(populate_by_name=True)
+
+    as_path: str = Field(serialization_alias="asPath", validation_alias="asPath")
 
 
 class ASPathList(PolicyListBase):
