+++ catalystwan/models/policy/list/data_prefix.py
@@ -3,8 +3,15 @@
 from ipaddress import IPv4Network
 from typing import List, Literal
 
-from catalystwan.models.policy.lists_entries import DataPrefixListEntry
+from pydantic import BaseModel, ConfigDict, Field
+
 from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
+
+
+class DataPrefixListEntry(BaseModel):
+    model_config = ConfigDict(populate_by_name=True)
+
+    ip_prefix: IPv4Network = Field(serialization_alias="ipPrefix", validation_alias="ipPrefix")
 
 
 class DataPrefixList(PolicyListBase):
