+++ catalystwan/models/policy/list/ips_signature.py
@@ -2,8 +2,16 @@
 
 from typing import List, Literal
 
-from catalystwan.models.policy.lists_entries import IPSSignatureListEntry
+from pydantic import BaseModel, ConfigDict, Field
+
 from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
+
+
+class IPSSignatureListEntry(BaseModel):
+    model_config = ConfigDict(populate_by_name=True)
+
+    generator_id: str = Field(serialization_alias="generatorId", validation_alias="generatorId")
+    signature_id: str = Field(serialization_alias="signatureId", validation_alias="signatureId")
 
 
 class IPSSignatureList(PolicyListBase):
