+++ catalystwan/models/policy/list/policer.py
@@ -2,8 +2,23 @@
 
 from typing import List, Literal
 
-from catalystwan.models.policy.lists_entries import PolicerExceedAction, PolicerListEntry
+from pydantic import BaseModel, ConfigDict, Field
+
+from catalystwan.models.common import IntStr
 from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
+
+PolicerExceedAction = Literal[
+    "drop",
+    "remark",
+]
+
+
+class PolicerListEntry(BaseModel):
+    model_config = ConfigDict(populate_by_name=True)
+
+    burst: IntStr = Field(description="bytes", ge=15_000, le=10_000_000)
+    exceed: PolicerExceedAction = "drop"
+    rate: IntStr = Field(description="bps", ge=8, le=100_000_000_000)
 
 
 class PolicerList(PolicyListBase):
