+++ catalystwan/models/policy/list/preferred_color_group.py
@@ -2,9 +2,50 @@
 
 from typing import List, Literal, Optional, Set, Tuple
 
-from catalystwan.models.common import TLOCColor
-from catalystwan.models.policy.lists_entries import ColorGroupPreference, PathPreference, PreferredColorGroupListEntry
+from pydantic import BaseModel, ConfigDict, Field, field_validator, model_validator
+
+from catalystwan.models.common import TLOCColor, str_as_str_list
 from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
+
+PathPreference = Literal[
+    "direct-path",
+    "multi-hop-path",
+    "all-paths",
+]
+
+
+class ColorGroupPreference(BaseModel):
+    model_config = ConfigDict(populate_by_name=True)
+
+    color_preference: Set[TLOCColor] = Field(serialization_alias="colorPreference", validation_alias="colorPreference")
+    path_preference: PathPreference = Field(serialization_alias="pathPreference", validation_alias="pathPreference")
+
+    _color_pref = field_validator("color_preference", mode="before")(str_as_str_list)
+
+    @staticmethod
+    def from_color_set_and_path(
+        color_preference: Set[TLOCColor], path_preference: PathPreference
+    ) -> "ColorGroupPreference":
+        return ColorGroupPreference(color_preference=color_preference, path_preference=path_preference)
+
+
+class PreferredColorGroupListEntry(BaseModel):
+    model_config = ConfigDict(populate_by_name=True)
+
+    primary_preference: ColorGroupPreference = Field(
+        serialization_alias="primaryPreference", validation_alias="primaryPreference"
+    )
+    secondary_preference: Optional[ColorGroupPreference] = Field(
+        default=None, serialization_alias="secondaryPreference", validation_alias="secondaryPreference"
+    )
+    tertiary_preference: Optional[ColorGroupPreference] = Field(
+        default=None, serialization_alias="tertiaryPreference", validation_alias="tertiaryPreference"
+    )
+
+    @model_validator(mode="after")
+    def check_optional_preferences_order(self):
+        assert not (self.secondary_preference is None and self.tertiary_preference is not None)
+        return self
 
 
 class PreferredColorGroupList(PolicyListBase):
