+++ catalystwan/models/policy/list/protocol_name.py
@@ -2,8 +2,15 @@
 
 from typing import List, Literal
 
-from catalystwan.models.policy.lists_entries import ProtocolNameListEntry
+from pydantic import BaseModel, ConfigDict, Field
+
 from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
+
+
+class ProtocolNameListEntry(BaseModel):
+    model_config = ConfigDict(populate_by_name=True)
+
+    protocol_name: str = Field(serialization_alias="protocolName", validation_alias="protocolName")
 
 
 class ProtocolNameList(PolicyListBase):
